+++ gnu/findutils/find/exec.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+set -e
+
+sudo find -type d -exec chmod 755 {} \;
+sudo find -type f -exec chmod 644 {} \;